Hadoop: dropNamespace should throw NamespaceNotEmptyException when not empty (#1200)

+++ core/src/main/java/org/apache/iceberg/hadoop/HadoopCatalog.java
@@ -42,6 +42,7 @@ import org.apache.iceberg.catalog.Namespace;
 import org.apache.iceberg.catalog.SupportsNamespaces;
 import org.apache.iceberg.catalog.TableIdentifier;
 import org.apache.iceberg.exceptions.AlreadyExistsException;
+import org.apache.iceberg.exceptions.NamespaceNotEmptyException;
 import org.apache.iceberg.exceptions.NoSuchNamespaceException;
 import org.apache.iceberg.exceptions.NoSuchTableException;
 import org.apache.iceberg.exceptions.RuntimeIOException;
@@ -276,8 +277,11 @@ public class HadoopCatalog extends BaseMetastoreCatalog implements Closeable, Su
     }
 
     try {
-      return fs.delete(nsPath, false /* recursive */);
+      if (fs.listStatusIterator(nsPath).hasNext()) {
+        throw new NamespaceNotEmptyException("Namespace " + namespace + " is not empty.");
+      }
 
+      return fs.delete(nsPath, false /* recursive */);
     } catch (IOException e) {
       throw new RuntimeIOException(e, "Namespace delete failed: %s", namespace);
     }
